requirements.txt
blinker==1.6.2
cfgv==3.3.1
click==8.1.6
colorama==0.4.6
distlib==0.3.6
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
identify==2.5.24
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.8.1
pre-commit==3.3.3
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyYAML==6.0
virtualenv==20.23.1
Werkzeug==2.3.6
